+++ test/cypress/integration/content-types.spec.js
@@ -1,7 +1,7 @@
 import {
   TEST_ID_PAGE_CONTAINER,
   TEST_ID_CONTENTTYPE_FORM,
-} from '../../../src/ui/test-const/contenttypes-complex-const';
+} from '../../../src/ui/test-const/content-types-const';
 
 describe('Content Types - Complex', () => {
   beforeEach(() => {
